backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # pin to avoid __about__ removal breaking passlib 1.x
python-multipart==0.0.9
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
woocommerce==3.0.0
python-dotenv==1.0.1
httpx==0.27.2
pymysql
psycopg2-binary
dotenv
requests==2.31.0
reportlab==4.0.9
arabic-reshaper==3.0.0
python-bidi==0.4.2
jdatetime==4.1.0
twilio==9.0.0
slowapi==0.1.9
structlog==24.1.0
aiofiles==23.2.1
Pillow==10.4.0
pytest==8.2.0
pytest-asyncio==0.23.6
websockets==13.0
redis